agents/identity/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "identity-agent"
version = "0.1.0"
description = "Identity Security Agent — Microsoft Entra ID posture assessment"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0,<3",
    "httpx>=0.27,<1",
    "pyyaml>=6.0,<7",
    "anthropic>=0.40,<1",
    "aiosqlite>=0.20,<1",
    "typer>=0.12,<1",
    "rich>=13,<14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "respx>=0.21,<1",
    "ruff>=0.5,<1",
]

[project.scripts]
identity-agent = "identity_agent.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/identity_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
